relocate duty_cycle decorators
