+++ tests/tests_rf/test_virt_network.py
@@ -52,7 +52,7 @@ SCHEMA_1 = {
 @pytest.fixture(autouse=True)
 def patches_for_tests(monkeypatch: pytest.MonkeyPatch):
     monkeypatch.setattr(
-        "ramses_tx.protocol._DBG_DISABLE_DUTY_CYCLE_LIMIT",
+        "ramses_tx.transport._DBG_DISABLE_DUTY_CYCLE_LIMIT",
         _DBG_DISABLE_DUTY_CYCLE_LIMIT,
     )
     monkeypatch.setattr(
